Create README.md
This repository is a lab assignment from my intro to C# course where I used different data types to provide course registration information for students enrolling in new classes.

Scenario: Students want to track the classes they are taking in the school. They want to track the schedule and they also want to maintain information about the courses they take, such as the course title, duration, credit hours, instructor’s name, and so on. Students will use the integral data types along with variables and constants to store this information in a computer program.

Objectives After completing this lab, you will be able to: • Describe the differences among the available data types. • Choose appropriate data type to represent real-world information. • Create and assign values to variables and use them in a program. • Create and assign values to constants and use them in a program.

Exercises 
• Selecting Data Types 
• Declaring and Using Variables for Numeric Types 
• Declaring and Using Variables for Textual Data Types 
• Declaring and Using Constants